investigate how to return seed and detach client that started schemachange
In previous checkin we manage to get comdb2sc to start schemachange in detached
mode and return seed to comdb2sc. It is able to do so because it connects directly
to the masternode and it starts the schemachange in detached mode.

We need to have ability for cdb2sql with ddl to do the same thing, but currently
the blocker for just setting is that replicant node servicing client will wait
for master node, and there is no way currently to return to the replicant node
seed. Furnthermore, there is complex logic for the thread on master to commit
the schema change, and that needs to be modified to run detached.