Allow unsigned int tunables to be assigned full range of values (#2814)
Currently setting an unsigned int tunable to values larger than LONG_MAX
results in failure. This change prepares us to properly assign values
up to ULONG_MAX.

We use strtoul() for all tunables that are NOT SIGNED, and strtol() for those
wich are SIGNED. The default assumed type of the tunables now will be unsigned
unless flag SIGNED is passed in.

To properly allow for values up to ULONG_MAX, a subsequent change will
be made to turn all the relevant tunables to unsigned int, and remove
the LONG_MAX check added in this checkin.

Signed-off-by: Adi Zaimi <azaimi@bloomberg.net>